% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\docType{data}
\name{alsroads_default_parameters}
\alias{alsroads_default_parameters}
\alias{alsroads_default_parameters2}
\title{Parameters}
\format{
An object of class \code{list} of length 7.

An object of class \code{list} of length 7.
}
\usage{
alsroads_default_parameters

alsroads_default_parameters2
}
\description{
Parameters for function \link{measure_road}. Keeping 'as is' is recommended.
}
\details{
\strong{extraction}: parameters used to extract the point cloud, process it into small sections
and compute profiles
\itemize{
\item \strong{road_max_len}: maximum size of a processed road. If longer than that it will be split in
chunks of equal sizes such as no one is more than this size.
\item \strong{road_max_width}: maximum width of a road for metrics measurements
\item \strong{road_buffer}: width of a buffer around the road for point-cloud extraction
\item \strong{section_length}: length of sections of road for metrics measurements
\item \strong{profile_resolution}: resolution of the profiles and DTMs for metrics measurements
}

\strong{constraint}: parameters to contraint the road relocation step
\itemize{
\item \strong{confidence} The confidence you have on the location of the reference roads. 1 means
that the road is 100\\% a ground truth. This will skip the relocation step. High values mean high
confidence that the reference road is correct and this will help the algorithm. Low values leave
more freedom to the algorithm but it becomes also more prone to errors. However this parameter is
not very sensitive.
}

\strong{embankments}: parameters to detect embankments in the terrain
\itemize{
\item \strong{min_slope}: slope (degrees) greater may initiate or terminate the detection of embankments
}

\strong{terrain}: parameters to detect road based on the terrain aspect (excluding embankment)
\itemize{
\item \strong{max_sd_ground_points}: point belonging on road have a low dispersion on Z.
maximum standard deviation to be considered as low dispersion
\item \strong{max_elevation_ground_points}: once normalized relatively to the potential road
all the points are expected to be at 0. This is the tolerance.
}

\strong{state}: parameters to evaluate the state of a section
\itemize{
\item \strong{percentage_veg_thresholds}:
\item \strong{drivable_width_thresholds}:
}
}
\examples{
str(alsroads_default_parameters)
}
\keyword{datasets}
