% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_post_prod.R
\name{check_road_differences}
\alias{check_road_differences}
\title{Check amplitude of differences between corrected and uncorrected roads}
\usage{
check_road_differences(roads, roads_ori, field)
}
\arguments{
\item{roads}{multiples lines (\code{sf} format). Corrected roads.}

\item{roads_ori}{multiples lines (\code{sf} format). Original non-corrected roads.}

\item{field}{character. Unique identifier field in both road datasets.}
}
\value{
data.frame with metrics about each roads.
}
\description{
Check amplitude of differences between corrected and uncorrected roads
}
\examples{
library(sf)

road <- system.file("extdata", "road_971487.gpkg", package="MFFProads")
road_ori <- st_read(road, "original", quiet = TRUE)
road_cor <- st_read(road, "corrected", quiet = TRUE)

df <- check_road_differences(road_cor, road_ori, "objectid")
df
}
