% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_road.R
\name{conductivity_from_bool}
\alias{conductivity_from_bool}
\title{Conductivity raster from boolean raster}
\usage{
conductivity_from_bool(x, w_max = 15)
}
\arguments{
\item{x}{raster (\code{raster} format). Boolean raster (1/0) where all 1 represent a pixel that might be categorized as a road.}

\item{w_max}{numeric. Maximum window size to use (start from 3x3). Must be an odd number over 3.}
}
\value{
raster (\code{raster} format) of conductivity with values ranging from 0 to 1.
}
\description{
Conductivity raster from boolean raster. Can be useful if one has a road raster generated
by classification algorithm and whant to extract centerlines from it. The moving window
will enhance the center of an object. In cases of roads (narrow elongated shape), the centerline
will becone more apparent.
}
