% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_road.R
\name{find_road_branches}
\alias{find_road_branches}
\title{Find potential secondary roads from main road}
\usage{
find_road_branches(road, conductivity)
}
\arguments{
\item{road}{line (\code{sf} format). Knowed road.}

\item{conductivity}{raster (\code{raster} format). Conductivity raster covering the road.}
}
\value{
lines (\code{sf} format) representing potential starting points of branching roads.
}
\description{
Tries to find secondary roads (branches) joining the main road using
a conductivity raster and the line geometry of the main road.
}
\examples{
library(sf)
library(raster)

road <- system.file("extdata", "drived_road.gpkg", package="MFFProads")
conductivity <- system.file("extdata", "drived_conductivity.tif", package="MFFProads")

road <- st_read(road, "drived_road", quiet = TRUE)
conductivity <- raster(conductivity)

road_branches <- find_road_branches(road, conductivity)

plot(conductivity, col = viridis::viridis(50))
plot(road, col = "green", lwd = 3, add = TRUE)
plot(road_branches, col = "red", add = TRUE)
}
