% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\docType{data}
\name{mffproads_default_parameters}
\alias{mffproads_default_parameters}
\title{Parameters}
\format{
An object of class \code{list} of length 4.
}
\usage{
mffproads_default_parameters
}
\description{
Parameters for function \link{measure_road}. Keeping 'as is' is recommended.
}
\details{
\strong{extraction}: parameters used to extract the point cloud, process it into small sections
and compute profiles
\itemize{
\item \strong{road_max_len}: maximum size of a processed road. If longer than that it will be split in
chunks of equal sizes such as no one is more than this size.
\item \strong{road_buffer}: width of the point-cloud extracted around the road
\item \strong{section_length}: length of sections of road for metrics measurements
\item \strong{profile_resolution}: resolution of the profiles and DTMs for metrics measurements
}

\strong{embankments}: parameters to detect embankments in the terrain
\itemize{
\item \strong{min_slope}: slope (degrees) greater may initiate or terminate the detection of embankments
}

\strong{terrain}: parameters to detect road based on the terrain aspect (excluding embankment)
\itemize{
\item \strong{max_sd_ground_points}: point belonging on road have a low dispersion on Z.
maximum standard deviation to be considered as low dispersion
\item \strong{max_elevation_ground_points}: once normalized relatively to the potential road
all the points are expected to be at 0. This is the tolerance.
}

\strong{state}: parameters to evaluate the state of a section
\itemize{
\item \strong{percentage_veg_thresholds}:
\item \strong{drivable_width_thresholds}:
}
}
\examples{
str(mffproads_default_parameters)
}
\keyword{datasets}
