% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conductivity.R
\name{rasterize_conductivity}
\alias{rasterize_conductivity}
\title{Compute the conductivity raster sigma}
\usage{
rasterize_conductivity(
  las,
  res = 1,
  dtm = NULL,
  param = mffproads_default_parameters,
  ...
)
}
\arguments{
\item{las}{an object of class LAS or LAScatalog from lidR}

\item{res}{numeric. Resolution of the output. If DTM is not NULL res is the resolution of the DTM}

\item{param}{a list of many parameters. See \link{mffproads_default_parameters}.}
}
\description{
Compute the conductivity raster sigma from lidar data according to Roussel et al. 2020
(see references)
}
\examples{
library(lidR)
library(raster)

dir  <- system.file("extdata", "", package="MFFProads")
dtm  <- system.file("extdata", "dtm_1m.tif", package="MFFProads")
ctg  <- readLAScatalog(dir)
dtm  <- raster(dtm)
las  <- readLAS
}
