% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_road.R
\name{tile_conductivity}
\alias{tile_conductivity}
\title{Generate and save to file tile of conductivity}
\usage{
tile_conductivity(bbox, dtm, ctg, outdir, buffer = 0)
}
\arguments{
\item{bbox}{named numeric vector. Bounding box vector with names \code{xmin}, \code{xmax}, \code{ymin}, \code{ymax}}

\item{dtm}{raster (\code{terra} format) or path to raster file. Digital terrain model covering \code{bbox}.}

\item{ctg}{\code{LAScatalog} covering \code{bbox}.}

\item{outdir}{character. Directory in which the produced tile will be saved.}

\item{buffer}{numeric (pixel unit). Buffer to be added outside of \code{bbox} when computing metrics. Won't affect extent of the produced tile.}
}
\value{
character representing filenames of generated tiles.
}
\description{
Generate and save to file tile of conductivity based on the extent
of an area of interest.
}
\examples{
library(sf)
library(terra)
library(lidR)
library(future)
library(future.apply)

dir <- system.file("extdata", "", package="MFFProads")
path_dtm <- system.file("extdata", "dtm_1m.tif", package="MFFProads")

ctg <- readLAScatalog(dir)
dtm <- rast(path_dtm)

# Define parameters for grid of tiles
# Internal buffer is only to mess things up a bit
aoi <- st_bbox(ctg) |>
  st_as_sfc() |>
  st_buffer(-14.7)
size <- 500  # (in distance unit)

# Generate grid of tiles
bbox <- st_bbox(aoi)
bbox[c("xmin","ymin")] <- floor(bbox[c("xmin","ymin")])

xlength <- bbox["xmax"] - bbox["xmin"]
bbox["xmax"] <- bbox["xmin"] + ceiling(xlength / xres(dtm)) * xres(dtm)
ylength <- bbox["ymax"] - bbox["ymin"]
bbox["ymax"] <- bbox["ymin"] + ceiling(ylength / yres(dtm)) * yres(dtm)

grid <- st_make_grid(bbox, size)

# Generate list of bounding boxes for parallelisation
bboxes <- lapply(grid, function(x)
{
  coords <- st_coordinates(x)
  bbox <- c(range(coords[,"X"]), range(coords[,"Y"]))
  names(bbox) <- c("xmin","xmax","ymin","ymax")
  bbox
 })

# Generate tiles in parallel
outdir <- tempdir()
buffer <- 5
plan(multisession)
filenames <- future_sapply(bboxes, tile_conductivity, path_dtm, ctg, outdir, buffer)
plan(sequential)

# Create VRT from tiles for future use (allow a similar workflow as using a LAScatalog for LAS files)
path_filenames <- file.path(outdir, filenames)
path_vrtfile <- file.path(outdir, "conductivity.vrt")
vrt(path_filenames, path_vrtfile)
}
