% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_road.R
\name{drive_road}
\alias{drive_road}
\title{Drive along an unknowned road}
\usage{
drive_road(start_line, conductivity, fov = 45, radius = 10, cost_max = NULL)
}
\arguments{
\item{start_line}{line (\code{sf} format)}

\item{conductivity}{raster (\code{terra} format)}

\item{fov}{numeric. Field of view (degrees) ahead of the search vector.}

\item{radius}{numeric (distance unit). Search radius used to find the next most probable point on the road.}

\item{cost_max}{numeric. Maximal cost allowed in the conductivity raster for point candidate to continue the search.
If no value is provided \code{radius * 6} will be.}
}
\value{
list, \code{line} being the road found (as \code{sfc}) and \code{cost} a numeric vector of cost at each invidual vertices of the line.
}
\description{
Attempt to drive along an unknowned road in a conductivity raster starting
only with a small piece of road segment pointing in the right direction.
}
\examples{
library(terra)
library(sf)

conductivity <- system.file("extdata", "drived_conductivity.tif", package = "MFFProads")
start_line <- system.file("extdata", "start_line.gpkg", package = "MFFProads")
conductivity <- rast(conductivity)
start_line <- st_read(start_line)

res <- drive_road(start_line, conductivity)

plot(conductivity, col = viridis::viridis(50))
plot(res$line, add = TRUE, col = "red", lwd = 2)
plot(start_line, add = TRUE, col = "green", lwd = 3)
plot(res$cost, type = "l")
}
