% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_road.R
\name{drive_road}
\alias{drive_road}
\title{Drive along an unknown road}
\usage{
drive_road(starting_road, conductivity, fov = 45, sightline = 10)
}
\arguments{
\item{starting_road}{line (\code{sf} format)}

\item{conductivity}{raster (\code{raster} format)}

\item{fov}{numeric. Field of view (degrees) ahead of the search vector.}

\item{sightline}{numeric (distance unit). Search distance used to find the next most probable
point on the road.}
}
\value{
list, \code{line} being the road found (as \code{sfc}) and \code{cost} a numeric vector
of cost at each invidual vertices of the line.
}
\description{
Drive along an unknown road in a conductivity raster starting only with a small piece
of road segment pointing in the right direction.
}
\examples{
library(raster)
library(sf)

conductivity <- system.file("extdata", "j53e_conductivity.tif", package = "ALSroads")
starting <- system.file("extdata", "j53e_drived_road.gpkg", package = "ALSroads")

conductivity <- raster(conductivity)
starting <- st_read(starting, "starting_road", quiet = TRUE)

res <- drive_road(starting, conductivity)

plot(conductivity, col = viridis::viridis(50))
plot(starting, add = TRUE, col = "green", lwd = 3)
plot(res, add = TRUE, col = "red", lwd = 2)
}
