% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mffp_road.R
\name{measure_road}
\alias{measure_road}
\alias{measure_roads}
\title{Align and measure a road from lidar data}
\usage{
measure_road(
  ctg,
  road,
  dtm,
  water = NULL,
  param = mffproads_default_parameters,
  ...
)

measure_roads(
  ctg,
  roads,
  dtm,
  water = NULL,
  param = mffproads_default_parameters
)
}
\arguments{
\item{ctg}{a non-normalized \link[lidR:LAScatalog-class]{LAScatalog} object from lidR package}

\item{road}{a single line (sf format) used as reference to search and measure road.}

\item{dtm}{RasterLayer storing the DTM with a resolution of at least of 1 m. Can be computed
with \link[lidR:grid_terrain]{grid_terrain}.}

\item{water}{a set of polygons (sf format) of water bodies. This is used to mask the water bodies
so they cannot be mistaken as a drivable surfaces. Not mandatory but can help. It also allows to
detect bridges above water.}

\item{param}{a list of many parameters. See \link{mffproads_default_parameters}.}

\item{...}{unused}

\item{roads}{multiples lines (sf format) used as reference to search and measure roads}
}
\value{
An sf object similar to the input with additional attributes and an updated geometry. If
the class is 3 or 4 the original geometry is preserved to prevent adding more error. The new attributes
are ROADWITH, DRIVABLEWIDTH, PERCABOVEROAD (percentage of points between 0.5 and 5 meter above the road)
SHOULDERS (average number of shoulders found), SINUOSITY, CONDUCTIVITY (conductivity per linear meters)
SCORE (a road state score) and CLASS (4 classes derived from the SCORE). See references
}
\description{
From a reference road (line), extracts the line with a buffer from the point cloud and computes
the exact positioning of the road (realignment). Then, using new the accurate  shape, computes
road metrics including its width, its drivable width, sinuosity as well as its state in four
classes (1. operating, 2. maybe operating, 3. maybe decommissioned, 4. decommissioned). The
function \link{st_snap_lines} allows to post-process the output to fix minor inaccuracies and
reconnect the roads than may no longer be connected because each road is processed independently.
}
\examples{
library(lidR)
library(sf)
library(raster)

dir  <- system.file("extdata", "", package="MFFProads")
road <- system.file("extdata", "road_971487.gpkg", package="MFFProads")
dtm  <- system.file("extdata", "dtm_1m.tif", package="MFFProads")
ctg  <- readLAScatalog(dir)
road <- st_read(road, quiet = TRUE)
dtm  <- raster(dtm)

# Voluntarily add more error to the road
crs <- st_crs(road)
st_geometry(road) <- st_geometry(road) + st_sfc(st_point(c(-8, 0)))
st_crs(road) <- crs

plot(dtm, col = gray.colors(25, 0, 1))
plot(ctg, add = TRUE)
plot(st_geometry(road), add = TRUE, col = "red")

res <- measure_road(ctg, road, dtm)
res
poly <- sf::st_buffer(res, res$ROADWIDTH/2)

plot(st_geometry(road), col = "red") # Inaccurate road track
plot(st_geometry(res), col = "blue", add = TRUE) # Corrected road track

domain <- "https://servicesmatriciels.mern.gouv.qc.ca:443"
path <- "/erdas-iws/ogc/wmts/Inventaire_Ecoforestier/Inventaire_Ecoforestier/default/"
tiles <- "GoogleMapsCompatibleExt2:epsg:3857/{z}/{y}/{x}.jpg"
url <- paste0(domain, path, tiles)
m = mapview::mapview(list(road, poly),
  layer.name = c("Inaccurate", "Corrected"),
  color = c("red", "blue"), map.type = "Esri.WorldImagery")
leaflet::addTiles(m@map, url)

\dontrun{
# DEBUG

options(MFFProads.debug.finding = TRUE)
options(MFFProads.debug.measuring = FALSE)
options(MFFProads.debug.metrics = FALSE)
res <- measure_road(ctg, road, dtm)

options(MFFProads.debug.finding = FALSE)
options(MFFProads.debug.measuring = TRUE)
options(MFFProads.debug.metrics = FALSE)
res <- measure_road(ctg, road, dtm)

options(MFFProads.debug.finding = FALSE)
options(MFFProads.debug.measuring = FALSE)
options(MFFProads.debug.metrics = TRUE)
res <- measure_road(ctg, road, dtm)
}
}
\references{
Roussel, J-R, Achim A (2022). Correction, update, and enhancement of vectorial forestry
road map using ALS data, a pathfinder and seven metrics.
}
