% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conductivity.R
\name{rasterize_conductivity}
\alias{rasterize_conductivity}
\title{Compute the conductivity raster sigma}
\usage{
rasterize_conductivity(
  las,
  dtm = NULL,
  param = alsroads_default_parameters,
  ...
)
}
\arguments{
\item{las}{an object of class LAS or LAScatalog from lidR}

\item{dtm}{RasterLayer. If NULL is provided a DTM is computed on the fly. But if a DTM is already
available it can be given to the function.}

\item{param}{a list of many parameters. See \link{alsroads_default_parameters}.}

\item{...}{ignored}
}
\value{
a RasterLayer or SpatRaster
}
\description{
Compute the conductivity raster sigma from lidar data according to Roussel et al. 2020
(see references)
}
\examples{
library(lidR)
library(raster)
dir  <- system.file("extdata", "", package="ALSroads")
dtm  <- system.file("extdata", "j5gr_dtm.tif", package="ALSroads")
ctg  <- readLAScatalog(dir)
dtm  <- raster(dtm)
las  <- readLAS(ctg$filename[1])

sigma <- rasterize_conductivity(las, dtm = dtm)
plot(sigma, col = viridis::inferno(30))

\donttest{
sigma <- rasterize_conductivity(ctg, dtm = dtm)
plot(sigma, col = viridis::viridis(30))
}
}
