% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snap_tools.R
\name{st_snap_lines}
\alias{st_snap_lines}
\title{Snap road endings}
\usage{
st_snap_lines(roads, ref, tolerance = 30, field = NULL)
}
\arguments{
\item{roads}{multiples lines (\code{sf} format). Corrected but unconnected roads.}

\item{ref}{multiples lines (\code{sf} format). Original non-corrected but connected roads. Can be
missing.}

\item{tolerance}{numeric (distance unit). Tolerance value used to snap the road endings.}

\item{field}{character. Unique identifier field in both road datasets. Relevant only is 'ref' is
not missing.}
}
\value{
The same object as \code{roads} but with corrected ending such that roads are connected.
}
\description{
Snap road endings together, with and without prior knowledge on road connections, to post-process
and correct inaccuracies generated by \link{measure_roads} and ensure getting topologically valid
network. If argument 'ref' is missing the method is very basic. The methods using a reference map
is more advanced.
}
