% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/line_tools.R
\name{st_snap_lines}
\alias{st_snap_lines}
\title{Snap road endings}
\usage{
st_snap_lines(roads, tolerance = 8)
}
\arguments{
\item{roads}{multiples lines (sf format)}

\item{tolerance}{numeric. Tolerance value in distance unit used to snap the road endings}
}
\value{
The same object provided in input but with corrected ending such as roads are connected.
}
\description{
Snap road endings together to post-process and correct inaccuracies generated by \link{measure_roads}
and ensure getting topologically valid network.
}
